Refactor : Lottos 일급 컬렉션으로 변경
