+++ src/main/java/lotto/model/Lottos.java
@@ -1,22 +1,34 @@
 package lotto.model;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class WinningCalculator {
-
+public class Lottos {
     private static final int DEFAULT_VALUE = 0;
+    private final List<Lotto> lottos;
 
-    public WinningResult calculate(final WinningLotto winningLotto,
-        final PurChasedLotto purChasedLotto) {
+    public Lottos(final List<Lotto> lottos) {
+        this.lottos = lottos;
+    }
+
+    public int getLottoCount() {
+        return this.lottos.size();
+    }
+
+    public List<Lotto> getLottos() {
+        return new ArrayList<>(lottos);
+    }
+
+    public Map<Rank, Integer> calculateMatchLotto(final WinningLotto winningLotto) {
         Map<Rank, Integer> result = initializeRankMap();
-        List<Lotto> lottos = purChasedLotto.getLottos();
         for (Lotto lotto : lottos) {
             Rank rank = winningLotto.compare(lotto);
             result.put(rank, result.get(rank) + 1);
         }
-        return new WinningResult(result);
+        return Collections.unmodifiableMap(result);
     }
 
     private Map<Rank, Integer> initializeRankMap() {
