+++ src/test/java/lotto/model/LottoTest.java
@@ -18,26 +18,26 @@ public class LottoTest {
 
     static Stream<Arguments> createLottoSuccess() {
         return Stream.of(
-                arguments(List.of(1, 2, 3, 4, 5, 6)),
-                arguments(List.of(11, 2, 3, 4, 5, 6)),
-                arguments(List.of(1, 22, 3, 44, 5, 6)),
-                arguments(List.of(1, 2, 33, 4, 15, 6))
+            arguments(List.of(1, 2, 3, 4, 5, 6)),
+            arguments(List.of(11, 2, 3, 4, 5, 6)),
+            arguments(List.of(1, 22, 3, 44, 5, 6)),
+            arguments(List.of(1, 2, 33, 4, 15, 6))
         );
     }
 
     static Stream<Arguments> createLottoLengthFail() {
         return Stream.of(
-                arguments(List.of(1, 2, 3, 4, 5, 6, 7)),
-                arguments(List.of(11, 2, 3, 4, 5)),
-                arguments(List.of())
+            arguments(List.of(1, 2, 3, 4, 5, 6, 7)),
+            arguments(List.of(11, 2, 3, 4, 5)),
+            arguments(List.of())
         );
     }
 
     static Stream<Arguments> createLottoDuplicateFail() {
         return Stream.of(
-                arguments(List.of(1, 1, 3, 4, 5, 6)),
-                arguments(List.of(2, 2, 2, 4, 5, 5)),
-                arguments(List.of(1, 1, 1, 1, 1, 1))
+            arguments(List.of(1, 1, 3, 4, 5, 6)),
+            arguments(List.of(2, 2, 2, 4, 5, 5)),
+            arguments(List.of(1, 1, 1, 1, 1, 1))
         );
     }
 
@@ -46,7 +46,7 @@ public class LottoTest {
     @DisplayName("여러 로또 넘버가 중복되지 않으면 정상")
     void 로또_중복(List<Integer> numbers) {
         assertThatCode(() -> new Lotto(createLottoNumbers(numbers)))
-                .doesNotThrowAnyException();
+            .doesNotThrowAnyException();
     }
 
     @ParameterizedTest
@@ -54,8 +54,8 @@ public class LottoTest {
     @DisplayName("여러 로또 넘버가 중복되면 예외를 던진다.")
     void 로또_중복_예외(List<Integer> numbers) {
         assertThatCode(() -> new Lotto(createLottoNumbers(numbers)))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("로또 넘버는 중복되면 안됩니다.");
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessage("로또 넘버는 중복되면 안됩니다.");
     }
 
     @ParameterizedTest
@@ -63,8 +63,8 @@ public class LottoTest {
     @DisplayName("로또 번호는 6개가 아니면 예외를 던진다")
     void 로또_번호_6개가_아니면_예외_던진다(List<Integer> numbers) {
         assertThatCode(() -> new Lotto(createLottoNumbers(numbers)))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("로또 번호는 6개이여야 합니다");
+            .isInstanceOf(IllegalArgumentException.class)
+            .hasMessage("로또 번호는 6개이여야 합니다");
     }
 
     @ParameterizedTest
@@ -72,11 +72,11 @@ public class LottoTest {
     @DisplayName("로또 번호는 6개가 아니면 정상 동작")
     void 로또_번호_6개이면_정상(List<Integer> numbers) {
         assertThatCode(() -> new Lotto(createLottoNumbers(numbers)))
-                .doesNotThrowAnyException();
+            .doesNotThrowAnyException();
     }
 
     @Test
-    @DisplayName("임의의 로또 번호가 로또에 포함되어 있는지 테스트")
+    @DisplayName("로또 번호가 로또에 포함되어 있는지 테스트")
     void 로또_포함() {
         Lotto lotto = new Lotto(createLottoNumbers(List.of(1, 2, 3, 4, 5, 6)));
 
@@ -85,7 +85,7 @@ public class LottoTest {
 
 
     @Test
-    @DisplayName("임의의 로또 번호가 로또에 포함되어 있는 않는 테스트")
+    @DisplayName("로또 번호가 로또에 포함되어 있는 않는 테스트")
     void 로또_포함_X() {
         Lotto lotto = new Lotto(createLottoNumbers(List.of(1, 2, 3, 4, 5, 6)));
 
@@ -94,7 +94,7 @@ public class LottoTest {
 
     private List<LottoNumber> createLottoNumbers(List<Integer> container) {
         return container.stream()
-                .map(LottoNumber::new)
-                .collect(Collectors.toList());
+            .map(LottoNumber::new)
+            .collect(Collectors.toList());
     }
 }
